Convert web3 plugin into provider
