+++ src/plugins/index.ts
@@ -1,20 +1,21 @@
 import { App } from 'vue';
-import Web3Plugin from '@/services/web3/web3.plugin';
 import store from '@/store';
 import i18n from '@/plugins/i18n';
 import router from '@/plugins/router';
 import blocknative from '@/plugins/blocknative';
 import VueVirtualScroller from 'vue3-virtual-scroller';
 import { VueQueryPlugin } from '@tanstack/vue-query';
+import { provideWeb3Plugin } from '@/providers/web3-plugin.provider';
 
 export function registerPlugins(app: App) {
+  //TODO: is this testable?? is this SSR?
+  provideWeb3Plugin(app);
   app
     .use(i18n)
     .use(router)
     .use(store)
     .use(VueQueryPlugin)
     .use(blocknative)
-    .use(Web3Plugin)
     .use(VueVirtualScroller);
   return app;
 }
