+++ src/services/web3/useWeb3.ts
@@ -1,8 +1,8 @@
 import { Network } from '@balancer-labs/sdk';
 import { Web3Provider } from '@ethersproject/providers';
-import debounce from 'lodash/debounce';
-import { computed, inject, reactive, ref } from 'vue';
 import { useQuery } from '@tanstack/vue-query';
+import debounce from 'lodash/debounce';
+import { computed, reactive, ref } from 'vue';
 
 import useNetwork from '@/composables/useNetwork';
 import QUERY_KEYS from '@/constants/queryKeys';
@@ -11,10 +11,10 @@ import {
   hasInjectedProvider,
 } from '@/services/web3/connectors/metamask/metamask.connector';
 
+import { useWeb3Plugin } from '@/providers/web3-plugin.provider';
 import { configService } from '../config/config.service';
 import { rpcProviderService } from '../rpc-provider/rpc-provider.service';
 import { switchToAppNetwork } from './utils/helpers';
-import { Web3Plugin, Web3ProviderSymbol } from './web3.plugin';
 import { web3Service } from './web3.service';
 
 /** STATE */
@@ -45,7 +45,7 @@ export default function useWeb3() {
     disconnectWallet,
     connectWallet,
     isBlocked,
-  } = inject(Web3ProviderSymbol) as Web3Plugin;
+  } = useWeb3Plugin();
   const appNetworkConfig = configService.network;
 
   const { networkId } = useNetwork();
